Changed Input Manager and Added Jumping
Improved the input management in order to allow for different characters to take inputs from different controllers, like one taking inputs from keyboard while another takes inputs from a controller. Updated the moving left and right functions with this, and then added a jumping function that only works if the player is grounded.